int is not enough

+++ 5/E.Bindian_Signalizing/signal.cpp
@@ -44,8 +44,8 @@ void calInfo(){
   }
 }
 
-int getRes(){
-  int res =0;
+long getRes(){
+  long res =0;
 
   for(int i=1; i<n; i++){
     if(h[i] < h[0]){
